refactor: Errors and util. methods locations
