+++ contracts/profit/src/error.rs
@@ -45,3 +45,9 @@ pub enum ContractError {
     #[error("[Profit] EmptyBalance. No profit to dispatch")]
     EmptyBalance {},
 }
+
+impl ContractError {
+    pub(crate) fn unsupported_operation(msg: &'static str) -> Self {
+        Self::UnsupportedOperation(String::from(msg))
+    }
+}
